separated HERO from the main program
courtesy Snehit, programs wiil me more spaced and clean
